Lecture 25 UMG
Add Buff Widget using world Sever time to replicate to clients, use Draw as a Box to display correctly a texture that  is stretch and when wrap that texture into a SizeBox when you want to propertly see it working.  Use of multicast delegate to replicate time and see the duration of the buff/debuff.